backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-adk
python-dotenv
python-pptx